Model changes

+++ src/main/java/com/bits/scalableservices/course/service/entity/Course.java
@@ -21,5 +21,5 @@ public class Course {
 	private String courseCode;
 	private int semester;
 	@Column(nullable = false)
-	private int departmentId;
+	private Long departmentId;
 }
